[FLIP > MAIN] Major security flaw removed - some styling changes (#43)
* fix: changed from perSie to perDu

* new: implemented userID context hook

* fix: fixes major security flaw while resetting password on profile page

* changed endpoint where password is send to
* implemented ResetPasswordDTO on frontend and updated on backend
* added new auth hook to reset password

* chore: form will be disabled after success

* chore: renamend "Genral" to "Global"

* chore: renmaned function according to file name

* refac: wrapped row values into typography component

* chore: changed text color

* Add cookie-parser package and update dependencies

* Fix logout function to clear cookie and send response

* refac: changed Link to IconButton component with useNavigate -1 as redirect

* chore: changed welcome text

* refac: changing password now requires new login

* chore: added 1 sec delay for logout after pw change

* fix: fixed double savingof hashed pw

* chore: question changed

+++ server/src/Modules/auth/auth.service.ts
@@ -174,7 +174,7 @@ export class AuthService {
 
 		const hashedPassword = await this.encryptionService.hashPassword(password);
 		const updatedUserID = await this.usersService.updateUserCredentials(userID, { password: hashedPassword });
-		await this.encryptionService.deleteToken(token);
+		if (token) await this.encryptionService.deleteToken(token);
 
 		this.eventEmitter.emit('mail.password-changed', new PasswordChangedEvent(user.id, user.email, user.first_name));
